fix: rss path

+++ plugins/generateRss.js
@@ -51,12 +51,14 @@ const createFeed = async (feed, HOST) => {
 
 const updateRssFile = () => {
   const rssFeedPath = path.resolve(__dirname, '..', 'dist', 'rss.xml')
-  const rssFeed = fs.readFileSync(rssFeedPath, 'utf8')
-  const rssFeedReplaced = rssFeed.replace(
-    '?>',
-    '?>\r\n<?xml-stylesheet href="/rss.xsl" type="text/xsl"?>'
-  )
-  fs.writeFileSync(rssFeedPath, rssFeedReplaced)
+  if (fs.existsSync(rssFeedPath)) {
+    const rssFeed = fs.readFileSync(rssFeedPath, 'utf8')
+    const rssFeedReplaced = rssFeed.replace(
+      '?>',
+      '?>\r\n<?xml-stylesheet href="/rss.xsl" type="text/xsl"?>'
+    )
+    fs.writeFileSync(rssFeedPath, rssFeedReplaced)
+  }
 }
 
 module.exports = {
